+++ Src/Commands/RemoveDataCommand.cs
@@ -11,7 +11,7 @@ namespace Ldbc.Commands
 	public sealed class RemoveDataCommand : Abstract
 	{
 		[Parameter(Position = 0, Mandatory = true)]
-		public string CollectionName { get; set; }
+		public ILiteCollection<BsonDocument> Collection { get; set; }
 
 		[Parameter(Position = 1, Mandatory = true)]
 		public string Filter { get; set; }
@@ -20,20 +20,12 @@ namespace Ldbc.Commands
 		public object Parameters { get; set; }
 
 		[Parameter]
-		public ILiteDatabase Database { get; set; }
-
-		[Parameter]
 		public SwitchParameter Result { get; set; }
 
 		protected override void BeginProcessing()
 		{
-			if (Database == null)
-				Database = ResolveDatabase();
-
-			var collection = Database.GetCollection(CollectionName);
-
 			var param = new InputParameters(Parameters);
-			var deletedCount = collection.DeleteMany(param.Expression(Filter));
+			var deletedCount = Collection.DeleteMany(param.Expression(Filter));
 			if (Result)
 				WriteObject(deletedCount);
 		}