+++ Release-Notes.md
@@ -1,5 +1,10 @@
 # Ldbc Release Notes
 
+## v0.1.0
+
+- New cmdlet `Get-LiteCollection`
+- `Add-LiteData`, `Get-LiteData`, `Remove-LiteData` use collection instances instead of names
+
 ## v0.0.1
 
 LiteDB 5.0.1